World generation active!
And uber simple test of world generation is now active. It generates a
layer of grass at height 0 and that it is. (I know not much, but it is
the start.)

Cloud height is configurable as it stands, and can use two settings, a
per world cloud height, and a client override. I have a network packet
up and running on client login, but I haven't decieded on the packet
structure for sending out info. Once we have that nailed down we can
start testing per world cloud height