Layer visulation
The world now generates the basics of a multi layerd world for the
purpose of visulation of the ocean/land noise layer. Also I did a little
refactoring and the language files and other assets can now be loaded.

The ant build script is now in place and working, and we now generate
chunks using forge's built in chunk which allows for generation higher
than 255 if wanted that at the moment I am limiting to 255.

Servers can now send per world cloud height information to the client
although the client doesn't handle this just yet.